timeout as float

+++ runner/job/run.py
@@ -118,7 +118,7 @@ grp = submit.add_argument_group("simulation mode (submit, background...)")
 #x = grp.add_mutually_exclusive_group()
 grp.add_argument('--max-workers', type=int, 
                  help="number of workers for parallel processing (need to be allocated, e.g. via sbatch) -- default to the number of runs")
-grp.add_argument('-t', '--timeout', type=int, default=31536000, help='timeout in seconds (default to %(default)s)')
+grp.add_argument('-t', '--timeout', type=float, default=31536000, help='timeout in seconds (default to %(default)s)')
 grp.add_argument('--shell', action='store_true',
                help='print output to terminal instead of log file, run sequentially, mostly useful for testing/debugging')
 grp.add_argument('--echo', action='store_true',
